OBJECTS-377 DateType is not being formatted on storage, resulting in unformatted Dates! refactoring fixes
