+++ smartcosmos-core/src/main/java/net/smartcosmos/objects/constraints/validators/ObjectsMetadataValidator.java
